pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "hikari-saru"
authors = [
    {name = "bison", email = "aliteralbison@protonmail.com"}
]
description = "A personal toolkit for building discord bots with hikari."
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.9"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License"
]
dependencies = [
    "hikari == 2.0.0.dev106",
    "hikari-lightbulb >= 2.2.0"
]

dynamic = ["version"]

[project.urls]
homepage = "https://github.com/a-bison/hikari-saru"
repository = "https://github.com/a-bison/hikari-saru"

[tool.setuptools]
packages = ["saru"]

[tool.setuptools.dynamic]
version = { attr = "saru.__version__"}

[tool.mypy]
disallow_untyped_decorators = true
disallow_incomplete_defs = true
disallow_untyped_defs = true

warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true

exclude = [
    'noxfile.py'
]